Move sync_flows_from_fs to setup.py
