+++ src/backend/base/langflow/initial_setup/setup.py
@@ -23,7 +23,7 @@ from emoji import demojize, purely_emoji
 from loguru import logger
 from sqlalchemy.exc import NoResultFound
 from sqlalchemy.orm import selectinload
-from sqlmodel import select
+from sqlmodel import col, select
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from langflow.base.constants import FIELD_FORMAT_ATTRIBUTES, NODE_FORMAT_ATTRIBUTES, ORJSON_OPTIONS
@@ -978,3 +978,36 @@ async def get_or_create_default_folder(session: AsyncSession, user_id: UUID) -> 
         msg = "Failed to get or create default folder"
         raise ValueError(msg) from e
     return FolderRead.model_validate(folder_obj, from_attributes=True)
+
+
+async def sync_flows_from_fs():
+    flow_mtimes = {}
+    while True:
+        try:
+            async with session_scope() as session:
+                stmt = select(Flow).where(col(Flow.fs_path).is_not(None))
+                flows = (await session.exec(stmt)).all()
+                for flow in flows:
+                    mtime = flow_mtimes.setdefault(flow.id, 0)
+                    path = anyio.Path(flow.fs_path)
+                    try:
+                        if await path.exists():
+                            new_mtime = (await path.stat()).st_mtime
+                            if new_mtime > mtime:
+                                update_data = orjson.loads(await path.read_text(encoding="utf-8"))
+                                try:
+                                    for field_name in ("name", "description", "data", "locked"):
+                                        if new_value := update_data.get(field_name):
+                                            setattr(flow, field_name, new_value)
+                                    if folder_id := update_data.get("folder_id"):
+                                        flow.folder_id = UUID(folder_id)
+                                    await session.commit()
+                                    await session.refresh(flow)
+                                except Exception:  # noqa: BLE001
+                                    logger.exception(f"Couldn't update flow {flow.id} in database from path {path}")
+                                flow_mtimes[flow.id] = new_mtime
+                    except Exception:  # noqa: BLE001
+                        logger.exception(f"Error while handling flow file {path}")
+        except Exception:  # noqa: BLE001
+            logger.exception("Error while syncing flows from database")
+        await asyncio.sleep(10)
